Refactor CORS tests for OpenAPI out into their own class
